Projeto/requirements.txt
annotated-types==0.7.0
dnspython==2.7.0
email_validator==2.2.0
idna==3.10
pydantic==2.10.6
pydantic_core==2.27.2
python-multipart==0.0.20
typing_extensions==4.12.2
pydantic~=2.10.2
uvicorn~=0.32.1
fastapi~=0.115.5
bcrypt~=4.2.1
SQLAlchemy~=2.0.36
fastapi-utils~=0.8.0
python-dotenv~=1.0.1
psycopg2-binary
starlette~=0.41.3
jinja2
python-jose~=3.3.0
passlib~=1.7.4
marshmallow==3.26.1
packaging==24.2
pip==24.0

